Remove unnecessary getter from TransactionSystem
